% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_esu.R
\name{choose.esu}
\alias{choose.esu}
\title{Choose ESU}
\usage{
choose.esu(esu.names)
}
\arguments{
\item{esu.names}{The ESUs/DPSs in the input csv file.}
}
\value{
A vector of ESU/DPS names.
}
\description{
Utility function to read in a esu choice from a list
}
\keyword{utility}
