% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mpg.R
\name{clean.mpg}
\alias{clean.mpg}
\title{Lighter clean up population names}
\usage{
clean.mpg(pops)
}
\arguments{
\item{pops}{A vector of the population names from the input csv file.}
}
\value{
A vector of cleaned population names.
}
\description{
This is similar to \code{\link{clean.pops}} but does lighter cleaning.
}
\author{
Eli Holmes, NOAA, Seattle, USA.  eli(dot)holmes(at)noaa(dot)gov
}
\keyword{utility}
