% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pops.R
\name{clean.pops}
\alias{clean.pops}
\title{Clean up population names}
\usage{
clean.pops(pops)
}
\arguments{
\item{pops}{A vector of the population names from the input csv file.}
}
\value{
A vector of cleaned population names.
}
\description{
Utility function to clean up the population names and strip run timing
in the population names in the input file.
Change this file if the abbreviations used need to be changed.
}
\author{
Eli Holmes, NOAA, Seattle, USA.  eli(dot)holmes(at)noaa(dot)gov
}
\keyword{utility}
