% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracwild_tables.R
\name{fracwild_table}
\alias{fracwild_table}
\title{Make average and geomean fraction wild tables}
\usage{
fracwild_table(
  wild,
  total,
  max.year = 2014,
  lenbands = 5,
  nbands = 5,
  type = c("mean", "geomean")
)
}
\arguments{
\item{wild}{The wild count as a matrix. It is up to the user where this comes from. It could come from the raw fracwild data times raw total data or come from the smoothed frac wild times smoothed total estimates.}

\item{total}{The total count as a matrix. See above notes on the wild count.}

\item{max.year}{The last year to use when constructing the bands}

\item{lenbands}{Number of years to average}

\item{nbands}{Number of averages to show.}

\item{type}{Type of average: mean or geomean.}
}
\value{
a data frame of the formatted table
}
\description{
Takes the wild and total data and makes tables of fraction wild via wild/total
and constructs averages for given number of years (typically 5 or 10).
The wild data might be shorter than total since if there is no fracwild info,
the wild data might not have that population. The total data will be subsetted
to only have the populations in the wild data.

The function wants the matrices where the rownames are the population names
and the colnames are the years. In the package output, these are call
total: \code{matdat.spawners} and wild: \code{matdat.wildspawners}.
}
\author{
Eli Holmes, NOAA, Seattle, USA.  eli(dot)holmes(at)noaa(dot)gov
}
\keyword{report}
