% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwctrends_options.R
\name{nwctrends.options}
\alias{nwctrends.options}
\title{NWCTrends options and plot variables}
\value{
Nothing is returned. The variables are set in the internal package environment.
}
\description{
Set up the plot default line widths, types and colors. Pass in as a list, such as \code{nwctrends.options = list(main.raw.pch = 1)}
with the variables that you wish to set. See details for the names and descriptions of the variables.
}
\details{
\code{main.} variables for the main plot with trends for each ESU
\itemize{
\item \code{main.title.cex} Size of the main ESU titles in the plot. Default is 1.5.
\item \code{main.poptitle.cex} Size of the population titles in the individual panels. Default is 1.
\item \code{main.ylabel.cex} Size of the y axis labels in the individual panels. Default is 0.8.
\item \code{main.total.lty} Line type for the smoothed total spawners line. Default is 1 (solid).
\item \code{main.total.lwd} Line width for the smoothed total spawners line. Default is 3 (thick).
\item \code{main.total.col} Line color for the smoothed total spawners line. Default is black.
\item \code{main.wild.lty} Line type for the smoothed wild spawners line. Default is 1 (solid).
\item \code{main.wild.lwd} Line width for the smoothed wild spawners line. Default is 1 (thin).
\item \code{main.wild.col} Line color for the smoothed wild spawners line. Default is red (#D44045) from the NMFS palette.
\item \code{main.raw.pch} Point type for the raw spawners data points. Default is 19 (solid circle).
\item \code{main.raw.col} Color for the raw spawners data points. Default is blue (#00467F) from the NMFS palette.
\item \code{main.ci.col} Color for the confidence polygon around the smoothed total spawners line. Default is "grey75".
\item \code{main.ci.border} Border for the confidence polygon around the smoothed total spawners line. Default is NA which is no border. See \code{\link{polygon}()} for the options.
\item \code{main.NAtotal.lty} Line type for the smoothed total spawners line  before the first data points. Default is 1 (solid).
\item \code{main.NAtotal.lwd} Line width for the smoothed total spawners line  before the first data points. Default is 3 (thick).
\item \code{main.NAtotal.col} Line color for the smoothed total spawners line  before the first data points. Default is "grey".
\item \code{main.NAci.col} Color for the confidence polygon around the smoothed total spawners line before the first data points. Default is "grey95".
\item \code{main.NAci.border} Border for the confidence polygon around the smoothed total spawners line  before the first data points. Default is NA which is no border. See \code{\link{polygon}()} for the options.
}

\code{fracwild.} variables for the fracwild plot
\itemize{
\item \code{fracwild.title.cex} Size of the main ESU titles in the plot. Default is 1.5.
\item \code{fracwild.poptitle.cex} Size of the population titles in the individual panels. Default is 1.
\item \code{fracwild.ylabel.cex} Size of the y axis labels in the individual panels. Default is 0.8.
\item \code{fracwild.lty} Line type for the smoothed fracwild line. Default is 1 (solid).
\item \code{fracwild.lwd} Line width for the smoothed fracwild line. Default is 2 (medium thick).
\item \code{fracwild.col} Line color for the smoothed fracwild line. Default is blue (#00467F) from the NMFS palette.
\item \code{fracwild.raw.pch} Point type for the fracwild raw data points. Default is 1 (open circle).
\item \code{fracwild.raw.col} Color for the fracwild raw data points. Default is black.
}

\code{prod.} variables for the productivity plot
\itemize{
\item \code{prod.title.cex} Size of the main ESU titles in the plot. Default is 1.5.
\item \code{prod.poptitle.cex} Size of the population titles in the individual panels. Default is 1.
\item \code{prod.ylabel.cex} Size of the y axis labels in the individual panels. Default is 0.8.
\item \code{prod.col.neg} Color of the negative productivity bars. Default is red (#D44045) from the NMFS palette.
\item \code{prod.col.pos} Color of the positive productivity bars. Default is green (#007934) from the NMFS palette.
}

\code{geomean.} variables for the geomeans plot
\itemize{
\item \code{geomean.title.cex} Size of the main titles in the plot. Default is 1.
\item \code{geomean.xaxis.cex} Size of the x axis tick labels in the individual panels. Default is 0.9.
\item \code{geomean.yaxis.cex} Size of the y axis tick labels in the individual panels. Default is 0.9.
\item \code{geomean.pch} Point type for the data points. Default is 19 (solid circle).
\item \code{geomean.cex} Point size for the data points. Default is 1.5.
\item \code{geomean.col.neg} Color of the negative data points. Default is red (#D44045) from the NMFS palette.
\item \code{geomean.col.pos} Color of the positive data points. Default is black.
\item \code{geomean.abline.lty} Line type for the horizontal reference lines. Default is 2 (dashed).
\item \code{geomean.abline.col} Line color for the horizontal reference lines. Default is "grey".
}

\code{trend.} variables for the trends plot showing the x-year (e.g. 15-year) trend values in a plot
\itemize{
\item \code{trend.title.cex} Size of the main title in the plot. Default is 1.
\item \code{trend.ylabel.cex} Size of the y axis label. Default is 1.
\item \code{trend.axis.cex} Size of the axis tick labels. Default is 1.
\item \code{trend.pch} Point type for the data points. Default is 19 (solid circle).
\item \code{trend.cex} Point size for the data points. Default is 1.5.
\item \code{trend.col.neg} Color of the negative data points. Default is red (#D44045) from the NMFS palette.
\item \code{trend.col.pos} Color of the positive data points. Default is black.
}
}
\seealso{
\code{NWCTrends_report()}
}
\author{
Eli Holmes, NOAA, Seattle, USA.  eli(dot)holmes(at)noaa(dot)gov
}
\keyword{utility}
