% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwctrends_palette.R
\docType{data}
\name{nwctrends.palette}
\alias{nwctrends.palette}
\title{Palette}
\format{
A list.
\describe{
\item{red}{#D44045}
\item{green}{#007934}
\item{blue}{#00467F}
\item{black}{#000000}
\item{white}{#FFFFFF}
}
}
\description{
Definition of the colors used in the plots. To change the
colors issue a command such as \code{nwctrends.palette$red <- "pink"}.
}
\examples{
data('nwctrends.palette')
}
