% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productivity_plot.R
\name{productivity_plot}
\alias{productivity_plot}
\title{Productivity plot}
\usage{
productivity_plot(
  esu,
  pops,
  total.fit,
  fracwild.fit,
  min.year = NULL,
  max.year = NULL,
  type = 1,
  lag = 4,
  nwctrends.palette = list(red = "#D44045", white = "#FFFFFF", green = "#007934", blue
    = "#00467F", black = "#000000")
)
}
\arguments{
\item{esu}{The name of the ESU}

\item{pops}{The population names that will be plotted.}

\item{total.fit}{total fit returned by \code{trend_fits()}}

\item{fracwild.fit}{fracwild fit returned by \code{trend_fits()}}

\item{min.year}{The x axis minimum. First year for numerator.}

\item{max.year}{The x axis maximum. Last year for numerator.}

\item{type}{The type of plot. Type 3: wild(t+1)/wild(t). Type 1: wild(t+lag)/total(t)}

\item{lag}{The number of years prior to use in the denominator, e.g. spawnwers(year-lag). Note not used if type=3.}

\item{nwctrends.palette}{The colors to use for the plots. The default is \code{list(red="#D44045", white="#FFFFFF", green="#007934", blue="#00467F", black="#000000")}}
}
\value{
A plot
}
\description{
This uses the smoothed spawner estimates and smoothed fracwild estimates to compute a
productivity metric. Type 3: \code{wild(t+1)/wild(t)}. Type 1: \code{wild(t+lag)/total(t)}, where \code{wild}
is smoothed total estimate times smoothed fracwild estimate and \code{total} is the smoothed total
estimate.

In the Viability Report, \code{type=1} and the lag is set to 3 or 4 (depending on species).
}
\author{
Eli Holmes, NOAA, Seattle, USA.  eli(dot)holmes(at)noaa(dot)gov
}
\keyword{report}
