% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_params.R
\name{report_params}
\alias{report_params}
\title{Parameters for ESU_report.Rmd}
\usage{
report_params(
  figdir,
  fits,
  datalist,
  this.esu = 1,
  plot.min.year = NULL,
  plot.max.year = NULL,
  output.type = "word",
  min.data.points = 5
)
}
\arguments{
\item{figdir}{Where output is saved.}

\item{datalist}{A list of input returned from \code{data_setup()}}

\item{this.esu}{If there there are multiple ESUs in the fitslist$fits object (a list), then this is the one to use. An integer.}

\item{plot.min.year}{Optional. The earliest year to use when plotting the data if different than the first year in the data set.}

\item{plot.max.year}{Optional. The last year to use when plotting the data if different than the last year in the data set.}

\item{output.type}{"html", "pdf", or "word" Format to produce the report in.}

\item{min.data.points}{The minimum data points to require from a population (for fitting and plotting).}

\item{fitslist}{A list of fits returned from \code{trend_fits()}}
}
\value{
A list of parameters.
}
\description{
This returns a list of the needed parameters for the report Rmd.
}
\author{
Eli Holmes, NOAA, Seattle, USA.  eli(dot)holmes(at)noaa(dot)gov
}
\keyword{internal}
