% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_15_tables.R
\name{trend_15_table}
\alias{trend_15_table}
\title{Create the trend tables}
\usage{
trend_15_table(
  pops,
  mpg,
  total.fit,
  fracwild.fit,
  year.ranges = list(1990:2005, 1999:2014),
  wild = TRUE,
  navalue = " "
)
}
\arguments{
\item{pops}{which populations to include in the table}

\item{mpg}{Population group. Shown in the table.}

\item{total.fit}{The matrix of total spawner estimates}

\item{fracwild.fit}{The matrix of fraction wild associated with each total row.}

\item{year.ranges}{The columns of years.}

\item{wild}{Show smoothed wild or smoothed total.}

\item{navalue}{Value to use for NAs in the table. Default is a blank.}
}
\value{
A data frames with the estimates trend for each year range in a different column.
}
\description{
Create the tables with the trends for different time periods using the smoothed
spawner estimates. The trend is the slope of a linear regression of the log spawner counts versus year. Despite the name of the function, the range of years need not be 15 years.
The years to show are specified
by list \code{year.ranges}. The ranges are specified as \code{begin.year:end.year}, for example
\code{1990:2005}. \code{year.ranges} can be padded into the NWCTrends_report() call by passing in
trend.table.control as list.  For example \code{list(year.ranges=list(1990:2000,2000:2010))}.
}
