% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{data_setup}
\alias{data_setup}
\title{Read in the inputfile}
\usage{
data_setup(inputfile, min.year, max.year)
}
\arguments{
\item{inputfile}{.csv file. See demofiles for the proper format.}

\item{min.year}{The minumum year for the returned data frames.}

\item{max.year}{The maximum year for the returned data frames.}
}
\value{
A list with four items:
\describe{
  \item{dat}{The raw data for the selected ESUs.}
  \item{matdat.spawners}{A matrix of the total spawners with NAs for missing years. Each column is a year from min.year to max.year and each row is a population.}
  \item{matdat.wildspawners}{A matrix of the the wildspawners using the fracwild data if included. NAs for years with either missing fracwild or missing spawner count. Each column is a year from min.year to max.year and each row is a population.}
  \item{metadat}{A data.frame with all the metadata for each population: name = population name, ESU = ESU name, Species, Run = runtiming for population, PopGroup = name of the Major Population Group (within ESU).}
}
}
\description{
Read in the csv inputfile and create the data frames and matrices needed for 
the fitting, plots and tables: matdat.spawners, matdat.wildspawners, and metadata. 
Some clean up of names and runtiming is done.
}
\details{
NAs are specified with -99, -99.00 or -99.0.
}
