% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{NWCTrends-package}
\alias{NWCTrends}
\alias{NWCTrends-package}
\title{NWFSC PNW Salmonid Viability Report Standardized Metrics}
\description{
This package creates the ESU/DPS tables and figures for the standardized metrics used in the
Northwest Fisheries Science Center's PNW Salmonid Viability Reports which support the 5-year
Status Review updates (prepared by WRO) for Pacific salmon and steelhead listed
under the Endangered Species Act: Pacific Northwest".
}
\details{
The 2015 NWFSC PNW Salmonid Viability Report can be viewed by typing
\code{RShowDoc("2015 Status Review Update",package="NWCTrends")}
at the command line. The report has a description of the
methods used for computing the smoothed trend lines and the status metrics. A pdf of the methods alone is also available by typing
\code{RShowDoc("Methods",package="NWCTrends")}
at the command line.

The default model used to fit the data is that used in the
2015 Status Update. This model uses information across the
populations in an ESU to estimate the process variance,
non-process variance (residuals
between smoothed fits and observed spawners), covariance in
process errors (good and bad year correlation). However it
allows each population to have a different trend.  This model is
specified as \code{ model=list(Z="identity", R="diagonal and equal", U="unequal", Q="equalvarcov") }.

The main function that creates the reports and plots is \link{NWCTrends_report}.
This calls \code{esu_report.Rmd} which makes the reports and figures.
The data are input as a csv file and a folder is created with the name of this
csv file. Within that folder is  a \verb{metadata.csv file} for the meta data from
the csv file and then there are individual files for each ESU/DPS within
the csv file.
}
\references{
Ford, M. J., K. Barnas, T. Cooney, L. G. Crozier, M. Diaz, J. J. Hard, E. E. Holmes, D. M. Holzer, R. G. Kope, P. W. Lawson, M. Liermann, J. M. Myers, M. Rowse, D. J. Teel, D. M. Van Doornik, T. C. Wainwright, L. A. Weitkamp, M. Williams. 2015. Status Review Update for Pacific Salmon and Steelhead Listed under the Endangered Species Act:  Pacific Northwest. National Marine Fisheries Service, Northwest Fisheries Science Center.
Available from the NWFSC Publications page.
}
\seealso{
\link{data_setup}, \link{trend_15_tables}, \link{trend_fits}, \link{geomean_tables}.
}
\author{
\strong{Maintainer}: Holmes Elizabeth \email{eli.holmes@noaa.gov}

}
\keyword{internal}
